Rename collider to volume. We can implement a volume lib for rust afterwards. we need just remove the dependency on piston::graphics and our Draw trait.
