+++ QuakeRL-Rust.files
@@ -1,3 +1,9 @@
+src/volume.rs
+src/app.rs
+src/game.rs
+src/main.rs
+src/player.rs
+src/settings.rs
 assets/block.png
 assets/ranger_avatar.png
 Cargo.toml
@@ -8,6 +14,9 @@ src/input.rs
 src/app.rs
 src/collider.rs
 src/game.rs
+src/input.rs
 src/main.rs
 src/player.rs
+src/render.rs
 src/settings.rs
+src/sprite.rs
